add user dev route - split user related dev route
